Ported changes to exercise module from 2.8, fixed a typo in the message files and added 2 messages. But there is a problem and i can't figure out why this keeps happening.
When a user has started an attempt in an exercise and time runs out, while he is neither in modules/exercise/index.php nor in modules/exercise/exercise_submit.php
then clicking on exercise name causes an error: $langLessonDoesNotExist. The url is the same, no typo but still haven't found why it causes this error.